{{cookiecutter.project_slug}}/requirements.txt
Django==2.2.12
django-cors-headers==3.2.1
{%- if cookiecutter.use_redis == 'y' %}
django-redis==4.12.1
{%- endif %}
{%- if cookiecutter.use_celery == 'y' %}
django-celery-beat==2.0.0
{%- endif %}
djangorestframework==3.11.0
drf-nested-routers==0.91
drf-yasg==1.17.1
flake8==3.8.1
psycopg2==2.8.5
Pillow==7.1.2
{%- if cookiecutter.use_celery == 'y' %}
celery==4.4.6
{%- endif %}
{%- if cookiecutter.use_sentry == 'y' %}
sentry-sdk==0.16.0
{%- endif %}